Rename markdown-generation file
